\name{rstudio-package}
\alias{rstudio-package}
\alias{rstudio}
\docType{package}
\title{
Tools and Utilities for RStudio
}
\description{
Tools, utility functions, and APIs for use within RStudio.
}
\details{
\tabular{ll}{
Package: \tab rstudio\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-17\cr
License: \tab AGPL-3\cr
}

Tools, utility functions, and APIs for use within RStudio.

}
\author{
RStudio

Maintainer: RStudio <info@rstudio.org>
}

\keyword{ package }
\seealso{
\code{\link[manipulate:manipulate-package]{manipulate}}
}

\examples{

}
