\name{viewData}
\alias{viewData}
\title{Invoke the RStudio Data Viewer}
\description{
Invoke the RStudio data viewer
}
\usage{
viewData(data, caption)
}

\arguments{
  \item{data}{An R object which can be converted to a data frame}
  \item{caption}{Display caption for data (defaults to name of object)}
}

\details{
  The passed R object is converted to a data frame and then displayed
}

