\name{previewRd}
\alias{previewRd}

\title{
Preview an Rd topic in the Help pane
}
\description{
Preview an Rd topic in the Help pane
}
\usage{
previewRd(rdFile)
}
\arguments{
  \item{rdFile}{Single element character vector containing the name of the Rd file to be displayed}
}

\examples{
\dontrun{
rstudio::previewRd("~/MyPackage/man/foo.Rd")
}
}